http://pmt.ccnmtl.columbia.edu/item/64614/
2) (c) The colored vertical bars of the scenario graphs are also a little lower than the seasonal carbon release values that they are suppose to represent. See example graph on wiki here: http://wiki.ccnmtl.columbia.edu/images/6/6c/Bargraphs.jpg


I delved into the bluff-src.js and found that the significant digits of the label where being truncated incorrectly. Thus, the bars were actually correct, but the y-axis labels were incorrect. (y-axis label was actually 26.7 rather than just 26, so the bar painted correctly at 26.2.) I added a little code to allow the user to adjust precision based on an attribute. And, removed a floor call in the _label function. This seems to have done the trick.



git-svn-id: http://svn.ccnmtl.columbia.edu/blackrock/trunk@18889 1f418930-7ff8-0310-b8ad-c653122473bc