convert conditional to a guard
The method as it was *will* throw a 500 error if it gets a request that
isn't a POST, since it doesn't return anything (let alone an HTTP
response of some kind).

I'm just refactoring, so I didn't fix that. But I did at least drop a
comment in there so someone else maybe will.