Connect env data to graph - PMT #107406
Because I'm importing the data from two CSV files instead of one,
Highcharts can't just import it automatically. I made the choice to
reformat the data in javascript instead of in the python data processor.
There is some cleanup in the code I'll need to do, as well as adding
some tests. Also, there's scaling issues with the env data, which we'll
address as Kevin and Matt start to see what the data looks like on the
graph.