render portal page when there is no content
Just addressing a small developer quality of life issue. If you spin up
a new instance without any content in the database, the portal page,
which is currently the default, just gives a 500 error.

With this, the portal page isn't terribly useful since there is no
content there by default, but at least you get a basic page with
navigation on the first load and know that your install is basically
working rather than a 500 error.