Short circuit data browse to display objects only
django-databrowse is a great library for browsing object models. However, Blackrock doesn't really need all this functionality and there's not an easy way to tell the library exactly what we want. There's also some older code in django-databrowse that is sparking 500 errors. This PR circumvents the issues by doing a little url routing before handing the action back to databrowse.